Extract script deployment to util
