+++ packages/plugins/plugin-script/src/util/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+export * from './deploy';
